// extensions/openhue/src/api/generated/src/models/GroupedLightGet.ts
export interface GroupedLightGet {
  /**
   * Type of the supported resources
   * @type {string}
   * @memberof GroupedLightGet
   */
  type?: string;
  /**
   * Unique identifier representing a specific resource instance
   * @type {string}
   * @memberof GroupedLightGet
   */
  id?: string;
  /**
   * Clip v1 resource identifier
   * @type {string}
   * @memberof GroupedLightGet
   */
  id_v1?: string;
  /**
   *
   * @type {ResourceIdentifier}
   * @memberof GroupedLightGet
   */
  owner?: ResourceIdentifier;
  /**
   *
   * @type {On}
   * @memberof GroupedLightGet
   */
  on?: On;
  /**
   *
   * @type {Dimming}
   * @memberof GroupedLightGet
   */
  dimming?: Dimming;
  /**
   *
   * @type {GroupedLightGetAllOfAlert}
   * @memberof GroupedLightGet
   */
  alert?: GroupedLightGetAllOfAlert;
  /**
   *
   * @type {GroupedLightGetAllOfSignaling}
   * @memberof GroupedLightGet
   */
  signaling?: GroupedLightGetAllOfSignaling;
}