// extensions/openhue/src/api/generated/src/models/SceneGet.ts
export interface SceneGet {
  /**
   *
   * @type {string}
   * @memberof SceneGet
   */
  type?: SceneGetTypeEnum;
  /**
   * Unique identifier representing a specific resource instance
   * @type {string}
   * @memberof SceneGet
   */
  id?: string;
  /**
   * Clip v1 resource identifier
   * @type {string}
   * @memberof SceneGet
   */
  id_v1?: string;
  /**
   *
   * @type {ResourceIdentifier}
   * @memberof SceneGet
   */
  owner?: ResourceIdentifier;
  /**
   * List of actions to be executed synchronously on recall
   * @type {Array<ActionGet>}
   * @memberof SceneGet
   */
  actions?: Array<ActionGet>;
  /**
   *
   * @type {SceneMetadata}
   * @memberof SceneGet
   */
  metadata?: SceneMetadata;
  /**
   *
   * @type {ResourceIdentifier}
   * @memberof SceneGet
   */
  group?: ResourceIdentifier;
  /**
   *
   * @type {ScenePalette}
   * @memberof SceneGet
   */
  palette?: ScenePalette;
  /**
   * Speed of dynamic palette for this scene
   * @type {number}
   * @memberof SceneGet
   */
  speed?: number;
  /**
   * Indicates whether to automatically start the scene dynamically on active recall
   * @type {boolean}
   * @memberof SceneGet
   */
  auto_dynamic?: boolean;
  /**
   *
   * @type {SceneGetAllOfStatus}
   * @memberof SceneGet
   */
  status?: SceneGetAllOfStatus;
}

/**
 * @export
 */
export const SceneGetTypeEnum = {
  scene: "scene",
} as const;
export type SceneGetTypeEnum = (typeof SceneGetTypeEnum)[keyof typeof SceneGetTypeEnum];